refactor: remove redundant declarations and move context providers
